+++ cmd/opentelemetry-prometheus-sidecar/main.go
@@ -28,6 +28,7 @@ import (
 
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
+	"github.com/google/uuid"
 	"github.com/lightstep/opentelemetry-prometheus-sidecar/cmd/internal"
 	"github.com/lightstep/opentelemetry-prometheus-sidecar/config"
 	"github.com/lightstep/opentelemetry-prometheus-sidecar/health"
@@ -39,6 +40,7 @@ import (
 	"github.com/lightstep/opentelemetry-prometheus-sidecar/tail"
 	"github.com/lightstep/opentelemetry-prometheus-sidecar/telemetry"
 	"github.com/oklog/run"
+	"go.opentelemetry.io/otel/semconv"
 	"github.com/pkg/errors"
 	"github.com/prometheus/common/version"
 	"github.com/prometheus/prometheus/pkg/labels"
@@ -88,6 +90,9 @@ func Main() bool {
 	// environment variable to avoid recursion.
 	isSupervisor := !cfg.DisableSupervisor && os.Getenv(supervisorEnv) == ""
 
+	// Unique identifer for this process.
+	svcInstanceId := uuid.New().String()
+
 	// Configure logging and diagnostics.
 	logger := internal.NewLogger(cfg, isSupervisor)
 
@@ -96,6 +101,7 @@ func Main() bool {
 	telem := internal.StartTelemetry(
 		cfg,
 		"opentelemetry-prometheus-sidecar",
+		svcInstanceId,
 		isSupervisor,
 		logger,
 	)
@@ -183,7 +189,7 @@ func Main() bool {
 		queueManager,
 		cfg.OpenTelemetry.MetricsPrefix,
 		cfg.Prometheus.MaxPointAge.Duration,
-		labels.FromMap(cfg.Destination.Attributes),
+		createResourceLabels(svcInstanceId, cfg.Destination.Attributes),
 	)
 
 	// Start the admin server.
@@ -305,6 +311,11 @@ func parseFilters(logger log.Logger, filters []string) ([][]*labels.Matcher, err
 		matchers = append(matchers, m)
 	}
 	return matchers, nil
+}
+
+func createResourceLabels(svcInstanceId string, extraLabels map[string]string) labels.Labels {
+	extraLabels[string(semconv.ServiceInstanceIDKey)] = svcInstanceId
+	return labels.FromMap(extraLabels)
 }
 
 func selfTest(ctx context.Context, promURL *url.URL, scf otlp.StorageClientFactory, timeout time.Duration, logger log.Logger) error {